event bus gateway Phase 1
