+++ specs/src/openapi.json
@@ -42,6 +42,9 @@
     "/api/tokens/media": {
       "$ref": "paths/api_tokens_media.json"
     },
+    "/api/tokens/worker": {
+      "$ref": "paths/api_tokens_worker.json"
+    },
     "/api/users/me": {
       "$ref": "paths/api_users_me.json"
     },
